[Update] clean code grid card selector and [Add] simple dialog
